Migrate from gradle-cache-action to setup-gradle (#82)
burrunan/gradle-cache-action seems unmaintained, with the last release in August 2020.

gradle/actions/setup-gradle is the officially maintained by Gradle action that has been getting a lot of work done recently.

This also removes the build scans. We've never needed to use build scans in context with these jobs in the past. This could be changed in the future if necessary.

See deephaven/deephaven-core#5161